// src/generated/client/models/Person.ts
export interface Person {
    /**
     * Person ID in Forecast.
     * @type {number}
     * @memberof Person
     */
    id: number;
    /**
     * Person first name.
     * @type {string}
     * @memberof Person
     */
    firstName: string;
    /**
     * Person last name.
     * @type {string}
     * @memberof Person
     */
    lastName: string;
    /**
     * Person email.
     * @type {string}
     * @memberof Person
     */
    email: string;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    monday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    tuesday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    wednesday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    thursday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    friday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    saturday: number;
    /**
     * Persons expected working hours in minutes.
     * @type {number}
     * @memberof Person
     */
    sunday: number;
    /**
     * Defining is person active e.g. is employeed.
     * @type {boolean}
     * @memberof Person
     */
    active: boolean;
    /**
     * Persons language
     * @type {string}
     * @memberof Person
     */
    language?: string;
    /**
     * Start date of employment, String.
     * @type {string}
     * @memberof Person
     */
    startDate?: string;
    /**
     * Amount of unspent vacations from last year.
     * @type {number}
     * @memberof Person
     */
    unspentVacations: number;
    /**
     * Amount of spent vacations from last year.
     * @type {number}
     * @memberof Person
     */
    spentVacations: number;
    /**
     * Persons minimum billable rate
     * @type {number}
     * @memberof Person
     */
    minimumBillableRate: number;
    /**
     * PersonId in keycloak
     * @type {string}
     * @memberof Person
     */
    keycloakId?: string;
}