// src/generated/client/models/PersonTotalTime.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Timebank
 * Timebank API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Person's total time
 * @export
 * @interface PersonTotalTime
 */
export interface PersonTotalTime {
    /**
     * 
     * @type {number}
     * @memberof PersonTotalTime
     */
    personId?: number;
    /**
     * Difference between expected and logged time.
     * @type {number}
     * @memberof PersonTotalTime
     */
    balance: number;
    /**
     * Total amount logged time of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    logged: number;
    /**
     * Total amount of project time logged of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    loggedProjectTime: number;
    /**
     * Total amount expected time of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    expected: number;
    /**
     * Total amount of internal time of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    internalTime: number;
    /**
     * Total amount of billable project time of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    billableProjectTime: number;
    /**
     * Total amount of non billable project time of person.
     * @type {number}
     * @memberof PersonTotalTime
     */
    nonBillableProjectTime: number;
    /**
     * String identifying the relevant year/month/week for the requested period of time total.
     * @type {string}
     * @memberof PersonTotalTime
     */
    timePeriod?: string;
}

/**
 * Check if a given object implements the PersonTotalTime interface.
 */
export function instanceOfPersonTotalTime(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "balance" in value;
    isInstance = isInstance && "logged" in value;
    isInstance = isInstance && "loggedProjectTime" in value;
    isInstance = isInstance && "expected" in value;
    isInstance = isInstance && "internalTime" in value;
    isInstance = isInstance && "billableProjectTime" in value;
    isInstance = isInstance && "nonBillableProjectTime" in value;

    return isInstance;
}

export function PersonTotalTimeFromJSON(json: any): PersonTotalTime {
    return PersonTotalTimeFromJSONTyped(json, false);
}

export function PersonTotalTimeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonTotalTime {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'personId': !exists(json, 'personId') ? undefined : json['personId'],
        'balance': json['balance'],
        'logged': json['logged'],
        'loggedProjectTime': json['loggedProjectTime'],
        'expected': json['expected'],
        'internalTime': json['internalTime'],
        'billableProjectTime': json['billableProjectTime'],
        'nonBillableProjectTime': json['nonBillableProjectTime'],
        'timePeriod': !exists(json, 'timePeriod') ? undefined : json['timePeriod'],
    };
}

export function PersonTotalTimeToJSON(value?: PersonTotalTime | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'personId': value.personId,
        'balance': value.balance,
        'logged': value.logged,
        'loggedProjectTime': value.loggedProjectTime,
        'expected': value.expected,
        'internalTime': value.internalTime,
        'billableProjectTime': value.billableProjectTime,
        'nonBillableProjectTime': value.nonBillableProjectTime,
        'timePeriod': value.timePeriod,
    };
}